Fix cmake manager to not rely on undefined evaluation order.
Compiling with clang showed tons of test failures. The issue was
basically due to something like:

QStack<int> s;
foo(s.pop(), s.count());

The code relied on a right-to-left evaluation, but this is undefined
behavior and clang is perfectly fine with breaking this. Note that
commas are not synchronization points. To fix this, one has
to rewrite the above as

QStack<int> s;
int val = s.pop();
foo(val, s.count() + 1);