Support lib64 locations and also search packages based on PATH env.
When the global FIND_LIBRARY_USE_LIB64_PATHS property is set to
true (done in UnixPaths.cmake usually), then we also search for
packages in the lib64/ sub folders.

Furthermore, we increase compatibility to CMake by also searching
in folders based on the PATH environment variable. For pathes in there
ending on /bin or /sbin, we instead search in the parent directory,
just like CMake does (see documentation for find_package).

This hopefully fixes the unit tests on the build slaves, which depend
on these features as far as I can see.