Properly support CMAKE_{PREFIX,INCLUDE,LIBRARY}_PATH env vars.
These variables are "special", they come in two forms:

The direct variable, i.e. ${CMAKE_*_PATH}, can be set in cmake scripts
or by passing the variable explicitly to cmake when calling it via -D.

Alternatively, the user can define _additional_ variables of the same
name as environment variables. Previously, the code only supported
the former, not the latter which is now fixed by this commit.

Additionally, the CMAKE_PREFIX_PATH was previously looked up in the
cache, but that variable is never stored in the cache by cmake.

Maybe this was supposed to be CMAKE_INSTALL_PREFIX?

Anyhow, this test should bring us one more step towards running and
passing the unit tests on the build slaves, which uncover these issues
nicely with its non-standard setup.