Cleanup code and expect failure of test.
Note that this introduces a better pattern on how to set the type
of a declaration. Doing it directly will not properly update the
type of the declaration, when last/currentType is changed, as there
is no reference/pointer semantic between that and the indexed type
of the declaration.

Thus, only open the type and set it before you close the declaration.