Fixup implementation of FunctionExpression handling.
The type should be found in the expression visitor, and opened
by the declaration builder. Furthermore, we also need to create
the proper contexts for the FunctionExpression in the context builder,
which is now done by adding a function which does this for both the
FunctionDeclaration and FunctionExpression nodes.

This fixes the unit test.