Optimize: Get rid of DUContextDynamicData::m_localDeclarationsMutex.
We have our BKL, i.e. the DUChainLock. No need to pay double with
yet another mutex in between. Probably the same can be done for more
mutexes in the code.

In theory, the current code only needs mutexes for the ItemRepos
which might be accessed outside of the duchain. If you grep thoug,
you'll find that many more mutexes exist...