Optimize: Remove obsolete duchain locks and constify statics.
The locks where apparently historically required, but nowadays its
perfectly fine to operate on AbstractTypes which are not part of
the duchain (yet). This is the case for the expr. visitor, which
just creates new types. This pattern is used in other areas as well,
so I'm fairly confident I don't break things that way. I also success-
fully parsed my big session froms cratch with this patch applied.

Still, better safe than sorry, so I put it into master.

The constifying of the statics should be a no-brainer.