Port IncludePathComputer away from deprecated Url API.
Instead of introducing Path here though, we simply operate on QStrings
directly as is done already in lots of other places. This simplifies
the code and is fine here as we are just looking for local string
paths anyways. The additional API or memory savings of Path are not
important here, considering that the list of include paths is small
and temporary anyways. The overhead of converting to Path first would
not be worth it here.

With this patch in place, no Path conversions to KUrl take
place anymore when running KDevelop and parsing code.