EnvironmentWidget: Always display placeholder line
Implement a custom proxy model that just adds another line at the end of
the list. This line servers as a placeholder item. When setting data on
that placeholder item, the EnvironmentWidget instance is notified and
takes care of adding the actual value to the EnvironmentGroupModel
instance.

The new PlaceholderItemProxyModel should be useful for similar widgets
as well. For example, the BreakpointModel uses a similar approach,
but includes the placeholder item directly in the source QAIM. => Cleanup up
BreakPointModel and use this proxy to simplify code.

REVIEW: 114197