Added the possibility to filter Make Targets
Added the possibility to filter Make Targets on a custom makefile
project. This is needed because some makefile generators creates
a lot of boilerplate (I'm looking at you, qmake. ) that do not play
well with the KDevelop's project view ( more than 60 targets created
for my projects for instance. ). This makes KDevelop usable again.

Signed-off-by: Tomaz Canabrava <tcanabrava@kde.org>