Optimize: decrease size of ProjectFile by 8 bytes.
The project name can be looked up on demand, it is only required for
the actively selected file in the quickopen model.

As these structs are created in high order, saving 8 bytes here can
simply add up to quite a lot of memory saved. Furthermore, it speeds
things up as copying is faster for smaller structs.