Sort files in out-of-source build folders to the end of QuickOpen.
This removes the ugly clutter of ../../... paths to generated files
from uic etc. which are part of CMake targets. You may still want
to open them and can by inserting the proper filter into quickopen.

Note that performance regresses compared to the previous commit,
and is now back to the initial performance of this code which used
QMap<QString, ProjectFile> instead of the sorted QList.

The sharedurls branch and its Path API will speed it up in the future.