GrepDialog: Fix issues with search pattern
Before this patch:
When pressing Ctrl+Alt+F, the search pattern of the grep
dialog was always empty. This is quite annoying in case you just did a
search, then want to search again with the same pattern but in a
different location: You first have to scroll through the old search
patterns first, then select the location, then press "Search".

After this patch:
The search pattern field now has the same semantics as when doing a
search via simple Ctrl+F.
When pressing Ctrl+Alt+F: If the current selection in the document is
empty, directly use the previously entered pattern. If the selection is non-empty,
use that selected text as pattern.