Don't crash when the CMakeLists.txt structure contains 'gaps'.
With gaps, I mean folder structures such as this:

/CMakeLists.txt -> add_subdirectory(foo/bar)
/foo -> (no CMakeLists.txt)
/foo/bar/CMakeLists.txt

In this case, when we edited the last CMakeLists.txt, KDevelop would
crash as it cannot find a context for the direct parent folder.
Instead, we now recurse until we hit the project root. If no context
can be found, we fallback to initialize the project with the current
folder.

Note: This could be reproduced 100% of the cases when editing a
CMakeLists.txt in one of the sub folders of kdevplatform/language.

BUG: 327120