Optimize non-altering mutating operations.
If a mutating operation (set...) is called with an argument equal
to the current value, do not call the potential costly prepareWrite.
If we e.g. do this:

QualifiedIdentifier id;
id.setExplicitlyGlobal(false);

We would allocate a dynamic object even though this is equal to the
constant (and shared) default-constructed identifier.