Optimize: Use QSet<IndexedString> instead of ReferenceCountedStringSet
For runtime-only information this has much better performance and
reduces the amount of accesses to the serialization database at the
cost of slightly larger memory consumption.

I generally dislike the idea of the StorableSet and it is one of our
biggest hotspots currently. In future, this should be refactored,
probably by using e.g. QSet for deserialized data and "simple" vectors
of data on-disk.