Do not guess include paths when we any where returned by the manager.
The IncludePathComputer was run quite frequently and is very costly.
Especially for cmake projects, the build manager does a good job at
finding the include paths already.

Thus the up to 16s (for my KDev* session) spent in running make and
interpreting its result can be saved when we instead rely on the
result given by the builder.

If you suddenly encounter issues with include paths that used to work,
please report that on bugs.kde.org.

CCMAIL: kdevelop-devel@kde.org