Refactor IncludePath{Computer,Resolver} interaction.
We now still call the resolver's includeResolver.resolveIncludePath
as that is required to support custom include paths in the hidden
.kdev_include_paths files.

Instead, we disable the make-resolution when we get any paths from
the build manager.

Also get rid of the code which tries to be faster by looking at
the DUChain to find include paths. It's not faster. Also note how
the IncludePathResolver has a cache internally.