NinjaBuilder: Don't fetch targets recursively
My argumentation for this patch: *If* there are targets in any of the subdirectories,
these are marked as dependency of one the targets in the current
directory. => We don't need to pass those targets explicitely to the ninja call.

This fixes the issue of importing targets from subdirectories that are
not marked as dependency of one of the targets in the current directory.
C.f. comment #1 in 328364.

CCBUG: 328364