NinjaBuilder: Be more clever choosing sane targets
Example:
From the project view:
Try to build a directory which doesn't contain any targets, e.g. src/a/b/.
Assume that src/a/ contains the target 'a'.

Previous behavior:
Just "ninja" is run.
=> Everything is recompiled.

New behavior:
The ninja builder tries to find the next ancestor in the folder
hierarchy with a non-empty lists of targets.
=> "ninja a" is run. => Just src/a/ is compiled.