DebugController: Avoid opening non-existing files
If the debugger stops in poll() and the user has debug symbols for glibc
installed, the source file attached to the poll() function may point to
an non-existing file. At least this is the case for Debian-based
systems. Here, poll() points to some file in /build/buildd/.

Example Backtrace:
0 poll() at
/build/buildd/eglibc-2.17/io/../sysdeps/unix/syscall-template.S:81

This patch avoids opening a new file for syscall-template.S

Note that this patch isn't perfect, the real fix is to select the proper
line in the frame stack model on pause/stop. Committing this to the
stable branch to at least get rid off this annoying behavior.

Also see https://wiki.debian.org/buildd.