Reparse translation units when possible.
The ParseSession is now an IAstContainer which is stored in the
top-context. We then look for it in the parse job, and if available,
reparse the translation unit instead of parsing it anew.

This is much faster, a simple file which includes QString e.g. takes
ca. 400ms initially, and <5ms on subsequent runs.

The ParseSession will be deleted when the top context is not
referenced anymore, which should happen eventually for non-opened
files.