Fix issues in Breakpoint{Model}
Issue 1:
In BreakpointModel::load: new Breakpoint(...) eventually causes a call
to BreakpointModel::reportChange. This is problematic since at this
point the model doesn't know about the newly created breakpoint
=> reportChange notifies about an invalid item.

The KIdentityProxyModel on top is very unhappy about that and asserts.
Fixed by making the Breakpoint implicitly register at the model at
construction

Issue 2:
dataChanged() was emitted for columns that don't exist in the model.
Fixed by checking for columnCount() first.