Build imported contexts and associate uses of declarations therein.
After parsing the file, we visit its inclusion stack and figure
out which files are not yet cached in the duchain.

Then, when visiting the translation unit to build the DUChain cache,
we ignore any contents from files which are already cached.

For all others we also fill the DUChain cache. Analogously, we can
now associate the uses of imported declaratinos properly.

Open questions are probably the following:

- in an import structure like A -> B -> C, where only C requires an
update, we probably have to update B and A as well.
- when in the above, after A was parsed, B is updated, we probably
loose the uses since we do not actually "update" anything but
create stuff from scratch every time.
- more stuff?