Save memory by disposing the TU for non-opened files.
We only want to cache the parse session and the TU therein in order
to quickly reparse opened documents. Files which are not (yet)
opened we can parse anew when they are changed and then cache.

Eventually we might also want to dispose the cache when a document
is closed.