Prefer non-empty when computing include paths.
In some CMake projects a file may be listed as a source file in a
'real' target but also listed as a source to a 'custom' target.
Custom targets don't handle includes, so depending on the order of
evaluation when computing include paths, one may or may not get any
includes.

This change causes include path computation to prefer file items in
targets over those not in targets (as was the case before) and also
prefer file items in targets with includes over those without
includes.

REVIEW: 115057