VCS annotations: use proper colors for the foreground + background
Previously, the code selected a background color at random, used the
default foreground color, and made the bar partially transparent.
That made it look bright on bright themes and dark on dark themes but the
text contrast was very bad for dark themes.
Now it actually generates bright / dark opaque colors depending on the
theme, and also uses a bright / dark foreground color (for the text)
depending on the theme.

Needs latest kate, setting the foreground color was broken there.