Fix regression: When everything fails, try to get old include paths.
System headers or 3rdparty library headers are not part of any project
and thus all other mechanisms to find include paths for these files
fail. Thus it is very important to keep the fallback mechanism which
looks for include paths of an old parsed version of the file.
This way, we at least get some include paths for such files instead
of none at all which breaks many things in funny ways.