New declarative based assistant.
REVIEW:115196
This patch implements a new widget for the code assistant,
which is more flexible in what it looks like and how it behaves.
It is placed inside the editor widget instead of at the bottom,
to avoid covering openin toolviews. It adjusts to the user's
editor colors. By default, it's at the top of the view, since
I think it will cover interesting code less often; when you
scroll to the very top, it jumps down to avoid covering text
at the top.
Showing the widget is slightly animated; but just slightly
to avoid being too disrupting. Sames goes for the colors, they
are deliberately just slightly different from the colors used
in the editor to avoid blinking.