Start porting away from KTextEditor::Document::activeView()
It's gone, so we need alternatives, for now we can find them in
idocument.h and idocumentcontroller.h.
We have IDocumentController::activeTextView() that will return us the text
view of the current text document in case it actually is a text document.
Also, we have IDocument::activeTextView() that will also return a
KTextEditor::View in case it is a text document.