assistant: Prevent QGraphicsView from raising a window activation event
Different from QToolBar (which the old assistant inherited from),
QGraphicsView raises a WindowActivate event when it is shown. There is a
flag (WA_ShowWithoutActivating) which *should* prevent that but it
does not work. Thus, we re-implement QGraphicsView's ViewportEvent and
take the Show event away from it.

This is an ugly hack. However, with Qt5, QDeclarativeView goes away
anyways; so it is guaranteed to go away in at most half a year,
when I port this code to KDevelop 5. Until then, it's not going to
break, since Qt4 will not change.

CCMAIL:mail@milianw.de