Quick open filtering: do path fuzzy matching only on file name
Matching the whole path generated lots and lots of unrelated results,
especially for shorter queries. And because it is inherently hard to sort
that kind of fuzzy matches, they couldn't even be sorted by match quality.
That was overall not beneficial to QuickOpen's usefulness in practice.
Thus, the next attempt is to only do the fuzzy matching on the last
path segment; that still allows things like "btbrowser" for
"backtracebrowser.cpp" or "cml" for "CMakeLists.txt", but "hello" no longer
matches "/home/user/log/foo.log" or so.
CCMAIL:kevin@kfunk.org