don't start php with -e option
Accoding to php docs it should 'Activate the extended information mode, to be used by a debugger/profiler.'
However if using xdebug as debugger/profiler this option is not required.

I actually don't have any idea what this option does, it causes the xdebug plugin to fail so remove it.