Cleanup and slightly optimize type creation.
Identified types can be created directly without looking at the
actual type of their declarations. Now we specialize for all
declarations which declare new IdentifiedTypes, bypassing the
makeType path.
The makeType path now can be sure that the optimized cases are not
benig passed to it, and can always findDeclaration as needed.