GDB: Handle missing 'original-location' gracefully
Tested with GDB as of version 7.7 (released in 2014, Jan).
We may not have the 'original-location' field in all cases.
This happened when "catch throw" was used in gdbinit.

TODO: Check what we can do in case we don't get a source location at
all. Currently we might end up having empty lines in the breakpoint
table.

@Vlas, please review