Fix background parser unit tests
Properly port to new ThreadWeaver API.

It is important to realize that the ThreadWeaver::QObjectDecorator
object needs to be enqueued in ThreadWeaver::Queue, *not* the ParseJob.
QObjectDecorator takes care of delegating the calls to the original ParseJob.

This code is super ugly (lots of dynamic_cast), but I fear we can't
really get rid off them.