Set manpage plugin to be disabled by default
I have heard from several people now that it's insanely slow, the same
experience I had. This is fault of both the broken man KIO and the manpage
plugin, which uses the man KIO synchroneously and blocks the UI while it
is working. Until somebody fixes it, let's not load it by default because
it is hard to figure out what causes the slowness if it happens for a user.