Add #include-suggestion to kdev-clang
REVIEW: 117098

Implements the old #include/forward decl support from cpplanguage with
kdev-clang as the base framework.

libclang does not actually enumerate various semantic errors, so the solution
is based on parsing the error message from clang. While it works for some
configurations it is subject to break on languages and on internal clang
changes. Hopefully this situation will change in the future.

See the cfe-dev list for my suggestion to add this feature to libclang:
http://lists.cs.uiuc.edu/pipermail/cfe-dev/2014-March/036036.html