Use DeclarationBuilder to build declarations before ExpressionVisitor tries to find their type
When Declaration wants to find the type of an expression, it starts by recursing
into it (as JS expressions can contain declarations) and then uses ExpressionVisitor
to find its type.

This commit rewrites ExpressionVisitor so that it stops recursing when it first
encounters a type. This way, launching ExpressionVisitor multiple times on
nested expressions does not takes too much time.

REVIEW: 117509