Recognize function expressions
Function expressions are expressions having a "function" type, for instance the
right side of:

var a = function() { return 3; }

For this example, DeclarationBuilder will create an anonymous function
declaration and set its type to "int" (because 3 is of type int).
ExpressionVisitor is then able to use this declaration in order to say
that the right side of the assignation is of type "function int ()".

REVIEW: 117516