Merge the types of all the return expressions of functions
If a function has several return statements whose expressions are of different
types, then these types are merged into an unsure type. This way, the user can
see all the possible return types of the function.

This commit duplicates the mergeTypes function of kdev-python. This duplicated
function will be replaced by a standard KDevelop function when it will become
available.

REVIEW: 117535