A base expression visitor class for dynamically-typed languages.
This is potentially useful for Python, PHP, JS and Ruby.
For documentation on what it does, see the docs in the header file.
It's not a lot of shared functionality, but it is a piece of shared /
suggested architecture which makes it a) easier to write new language
plugins because there is a clear path for solving this problem; and b)
easier to understand other language plugin's architecture, because they
all work similar.