Search and load qch files on startup in an specified directory.
Quick implementation for searching qch files in an specified
directory. The directory can be specified in the QtHelp Dialog.
KDevelop searches for the qch files on startup. The imported qch
files are NOT added to the manual documentation list. The default
icon is the qtlogo an the name is the basename of the file.

REVIEW: 117014