Kill the MissingIncludeCompletionModel and worker.
Most of its functionality was already copied in the CompletionContext.
Just removing it showed nearly no regressions in my tests. The stuff
that is required, I now added to other places to make it work.

Also, for the fun of it: Note how not only was the functionality
partially duplicated by the MissingIncludeModel and normal completion
context, no! The doSpecialProcessing method in the model was also
usually called twice, once via updateCompletionRange and once from
startWithExpression...

Additionally, I don't see the need for a special worker thread here
in the first place. Only if normal code completion fails do we want
the missing include completion items, no?

Since this code is not unit tested at all, please test this! I hope
I did not miss any magic functionality of the model...