Hack around the race condition in CppUtils::findIncludePaths.
This time, we ensure the DUChain is not locked while delegating
work to the foreground thread via the infamous ForegroundLock.

It seems to work. But I said that the last time as well, so please
test! Generally, this code is a mess and I'm desparately awaiting
the time where I can say: "Good riddance!" kdev-clang to the rescue!