Use the cursor extent from clang directly when creating contexts
The reason we didn't do this earlier is because this means that
declarations with a context (functions, classes, namespaces) are
then inside of their own context range, which causes a problem
when finding said declaration.
However, it turns out that such things as template params are
also sementically within the function's context, and so they
actually need to be included in the function context's range.
Finding the declaration isn't a complicated condition, and
since it actually makes the range building cheaper, it's a net
win anyhow.