Add a template specific project name regex
Review 1174707 allows a project template to use a more specific regex for their project name validation.

This patch adds a template specific regex for all C++ templates that KDevelop provides. A C++ project can start now with a-Z or a '_'. After the first character, a-Z, 0-9 and '_' can be used.

REVIEW: 116978