Ensure that there is always an openType() corresponding to a closeType()
DeclarationBuilder::visit(UiObjectDefinition*) must ensure that a type and
a declaration will always be opened, because the corresponding endVisit()
method will always try to close a them. openType() and openDeclaration() cannot
therefore appear in a conditional statement that has no else clause.

This fixes a crash when the user types "key:" without a value in a QML component.