Ensure that getDeclaration and the code-completion can find declarations independently of their position in a QML file
QML and Javascript are highly dynamic languages and most of their declarations
can be used before the declaration itself. Now that the parser does two parsing
passes, this commit ensures that the location of an use is completely ignored
and that the parser always sees the complete list of declarations in a context.