Make it possible to show multiple toolviews again.
This was broken by the recent refactoring. The first call to
showWidget had the proper value for forceGrouping, but note that
this function triggers a recursion as it sets the checked property
of the action.

Instead, we now use the raise property for that, but set it
in buttonPressed, outside of showWidget.