Don't put the declarations of a module in a namespace, but store them directly in the top context
A stack of booleans allows endVisit(UiObjectDefinition) to know if a declaration
has to be closed. Module {} declarations are now skipped (but their content is explored),
and all the import logic is greatly simplified (no need for namespace aliases).

This is mostly a cleanup commit: namespaces didn't add anything useful and
were very complex to handle.