Hack around missing include paths for Qt4 files.
Quite often, Qt projects fail to parse essential headers like for
QString and similar. This is *probably* because they fail to find
the files they include in turn, which have the odd structure of
using the qt module as well... So /usr/include/qt4/QtCore/bla.h fails
to find <QtCore/blub.h>.

We now hack around this issue by falling back to a local lookup for
blub.h which is then found properly.

I hope this resolves most of the issues of QString being interpreted
as a function pointer and all the macros like foreach etc. pp.
not being available properly.