Do not freeze when executing a job with a lot of output.
We group the scroll-to-bottom events now which greatly reduces the
CPU overhead. Finding the visual rect and/or scrolling to the bottom
of a view is expensive, as it requires layouting of all rows. Now,
we do this at most ~3times per second, which is much less resource
hungry and leaves room for the user to cancel a job.

BUG: 336267