Further development of adjust signature assistant
Improvements which make the assistant behave more consistently.

However, making it work in as many cases as possible will require work on the core of kdev-clang. This is because it runs into issues where a header file cached in a translation unit is 'stale' (when it has been changed and saved since the translation unit was parsed). This breaks location queries in the cached header file, and I don't think there's any good way to avoid such queries while trying to implement the signature adjustment functionallity. Thoughts?

REVIEW: 118552